misc/HtmlTestSuite/results/test-0604.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Fonts Module Level 3: parsing font-feature-settings with valid values</title>
<link href="https://www.w3.org/TR/css-fonts-3/#font-feature-settings-prop" rel="help"/>
<meta content="font-feature-settings supports the full grammar 'normal | &lt;feature-tag-value&gt; #'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_valid_value('font-feature-settings', 'normal');

// &lt;feature-tag-value&gt; = &lt;string&gt; [ &lt;integer&gt; | on | off ]?
test_valid_value('font-feature-settings', '"dlig" 1', '"dlig"');
test_valid_value('font-feature-settings', '"smcp" on', '"smcp"');
test_valid_value('font-feature-settings', "'c2sc'", '"c2sc"');
test_valid_value('font-feature-settings', '"liga" off', '"liga" 0');
test_valid_value('font-feature-settings', '"tnum", \'hist\'', '"tnum", "hist"');

test_valid_value('font-feature-settings', '"PKRN"');

test_valid_value('font-feature-settings', '"dlig" 1, "smcp" on, "dlig" 0', '"dlig", "smcp", "dlig" 0');
</script>
</body>
</html>
